dashboard_analytics/requirements.txt
# Dashboard Analytics Requirements

# Streamlit
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP Client (for API calls if needed)
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development
pytest==7.4.3
